refactor: evaluator 모듈 내부 로직을 다른 파일로 분리
