+++ frontend/src/modules/evaluator/index.ts
@@ -1,67 +1,10 @@
 import { createObserver, type Listener } from '@/utils/observer';
 
 import EvalWorker from './eval.worker?worker';
-import type { EvalMessage, EvalResult } from './types';
+import EvalTaskManager from './EvalTaskManager';
+import type { EvalMessage, EvalResult, TaskEndMessage } from './types';
 
 const TOTAL_WORKERS = 3;
-
-type Evaluator = {
-  isIdle: boolean;
-  worker: Worker;
-  currentTask: EvalMessage | null;
-};
-type TaskEndMessage = EvalResult & { task: EvalMessage | null };
-
-const taskEndNotifier = createObserver<TaskEndMessage>();
-
-const evalManager = {
-  queuedTasks: [] as EvalMessage[],
-  evaluators: [] as Evaluator[],
-  addEvaluator(evaluator: Evaluator) {
-    this.evaluators.push(evaluator);
-  },
-  queueTasks(tasks: EvalMessage[]) {
-    this.queuedTasks.push(...tasks);
-  },
-  isWorking() {
-    return this.queuedTasks.length > 0;
-  },
-  popTask() {
-    return this.queuedTasks.shift();
-  },
-  deployTask() {
-    const restingEvaluator = this.evaluators.find((evaluator) => evaluator.isIdle);
-    if (!restingEvaluator) return;
-
-    const task = this.queuedTasks.shift();
-    if (!task) return;
-
-    restingEvaluator.isIdle = false;
-    restingEvaluator.currentTask = task;
-    restingEvaluator.worker.postMessage(task);
-
-    if (this.queuedTasks.length > 0) {
-      this.deployTask();
-    }
-  },
-  receiveTaskEnd({ result }: EvalResult, evaluator: Evaluator) {
-    taskEndNotifier.notify({
-      result,
-      task: evaluator.currentTask,
-    });
-
-    evaluator.isIdle = true;
-    evaluator.currentTask = null;
-    this.deployTask();
-  },
-};
-
-function init() {
-  for (let i = 0; i < TOTAL_WORKERS; i++) {
-    const worker = evaluatorFactory.createEvalWorker();
-    evalManager.addEvaluator(worker);
-  }
-}
 
 const evaluatorFactory = {
   createEvalWorker() {
@@ -80,6 +23,11 @@ const evaluatorFactory = {
     return evaluator;
   },
 };
+const taskEndNotifier = createObserver<TaskEndMessage>();
+const evalWorkers = Array(TOTAL_WORKERS)
+  .fill(undefined)
+  .map(() => evaluatorFactory.createEvalWorker());
+const evalManager = new EvalTaskManager(taskEndNotifier, evalWorkers);
 
 function safeEval(tasks: EvalMessage[]) {
   if (evalManager.isWorking()) {
@@ -93,8 +41,6 @@ function safeEval(tasks: EvalMessage[]) {
 function subscribe(listener: Listener<TaskEndMessage>) {
   taskEndNotifier.subscribe(listener);
 }
-
-init();
 
 export default {
   safeEval,
