+++ frontend/src/modules/evaluator/types.ts
@@ -8,3 +8,11 @@ export type EvalMessage = {
 export type EvalResult = {
   result: unknown;
 };
+
+export type Evaluator = {
+  isIdle: boolean;
+  worker: Worker;
+  currentTask: EvalMessage | null;
+};
+
+export type TaskEndMessage = EvalResult & { task: EvalMessage | null };
